python-service/requirements.txt
# gRPC dependencies
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4

# RabbitMQ client
pika==1.3.2

# Async support
asyncio==3.4.3

# Utilities
numpy==1.26.0
pyyaml==6.0.1

# Logging
python-json-logger==2.0.7